Let a replacement screen pop up, launch upon exit.
The replacement screen is, for the moment, the player options screen. (That's also the basis for the map options popup shown by the Full Of Resources script, which has been my inspiration.)

One problem solved is how to let the DLL determine that the Custom Game screen has been entered. Just some keeping track of calls to CvInitCore.

Another problem solved is how the replacement screen can launch the game. I don't think the necessary functions in the EXE are available to us. Conveniently, the launch button on the Custom Game screen seems to receive focus, reliably, when the replacement screen is closed. (More likely: It receives focus when the Custom Game screen is opened, and retains it.) So, all we need to do is simulate a press of the return key. A function for that is available in WinUser.h, which our precompiled header already includes via windows.h.

Next up: A large empty, maximized screen hiding the original Custom Game screen entirely. With a single widget for making some game setting. Remaining problem: Let the setting from the replacement screen have the last word over the Custom Game screen. Shouldn't be too difficult.